requirements.txt
anyio==4.3.0
asgiref==3.8.1
certifi==2024.2.2
Django==5.0.6
exceptiongroup==1.2.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
pillow==10.3.0
python-telegram-bot==21.1.1
sniffio==1.3.1
sqlparse==0.5.0
typing_extensions==4.11.0
